Adding initial provisoning code
When complete the 'provisioning' directory will be capable of deploying this application to a local development and remote production environment, including installing all necessary software, configuration files, databases and data files.

This can then be adpated by ICT for deploying the project to the BAS production cluster.

Currently, provisioning only goes as so far as provisioning a new database with PostGIS and relevant extensions enabled within a local development environment.

This will be extended to install GeoServer (with dependencies on Tomcat and Oracle Java), followed by further configuration to customise both aspects to the needs of the project.

This also includes standard files from within the Pristine base project, including editor configuration, licensing and a project README.