Adding java and tomcat basic support
Currently this only installs tomcat and configures a user for using the manager application.

Further work is needed to remove default applications and to deploy geoserver.

Adding tomcat has required an additional 'admin' account, which is part of the 'adm' group for editing config files and viewing logs. This account should be replaced by per-user accounts when the list of users is known.