Adding develpment environment using BAS development cluster
This adds a development environment based on the BAS development cluster.

This can then be used as the basis for providing a 'production-bas' environment which will be configured by ICT using Puppet.

Currently this uses a manual VM creation step and a static inventory, but all provisioning is then carried out by Ansible as normal. In time these manual aspects will be replaced by Terraform using vCentre.