Adding minimal docker support
Adds support for running a local instance of Map Engine in a Docker container.

This is currently a minimal implementation to allow high priority configuration to be made (see other issues). It consists of a multi-stage Dockerfile that:

1. builds static web assets (JS/CSS) using Grunt in a NodeJS container
2. builds Java application using Maven in a Java container as a WAR file
3. deploys WAR file into a Tomcat container

Currently the Maven build processs takes a significant amount of time due to large number of dependencies downloaded on each build. This can be improved in future by caching these dependencies using a Docker volume.

To run, the Map Engine application requires a PostGIS database and GeoServer instance. The former is provided through a separate PostGIS Docker container. Map Engine already includes a table setup SQL script, which is run on container startup.