Adding application properties profile for Docker
Map Engine (through Spring) uses profiles for configuration settings, such as database connection details.

Profiles are hierarchal with a base properties file defining defaults and additional instance specific files defining overrides for each instance (i.e. this instance is named ‘Foo’).

A profile is needed for Docker is needed to use the local database container and a remote GeoServer for authentication.

This Docker profile is minimal and uses environment variables as value placeholders to avoid recording variable or sensitive values.

Docker Compose will set these environment variables from an environment file `.env` which is excluded from this repository to prevent leaking secrets.

This file will need to be created as part of setting up this project locally. An example file `.env.example` is included to guide this (i.e. copy to `.env` and replace placeholder values).