Removing hard coded database user in setup script
The SQL script used to create required application tables assumed the presence of an `add` user, and will assign all tables it creates to this user.

This user doesn’t exist in a Docker environment (or logically in any environment other than the ADD instance). Therefore these statements have been removed, meaning tables will now be created by the Postgres user than runs the SQL script - which in Docker for example is the Postgres superuser.