Adding sample data to local GeoServer
Consists of data from the ADD data repository re-packaged as GeoPackages (with embedded styles). GeoPackages are used for future compatibility and to reduce the number of ancillary files.

These packages are added as vector layers in GeoServer into an ADD (for sea mask and Antarctic coastline layers) and a General workspace (for COMNAP facilities).

Note embedded samples are not currently read by GeoServer and so are created manually.

It’s expected these files will be replaced by official datasets in the future.