Refactoring Packer template
Splitting template into:

- stack (OS, Java, Tomcat, PostgreSQL, PostGIS)
- base (GeoServer, Web Map Engine)
- app (GeoServer data directory, Web Map Engine map configs)

This allows us to only rebuild the elements that have changed, which is frequently just the config or the application (‘app’ and ‘base’ templates). The ‘base’ template is intended for creating other instances, such as the Thwaites GIS.

Only the ‘stack’ template builds from an ISO, the others build from an existing OVA (created by stack) so they are much quicker to create, which should help with CI/CD.

As part of splitting these templates up some of the provisioning scripts have either been split or modified to ensure responsibilities are separated.

CI has been updated to take account of this new structure, with a one job to build the ‘base’ if the ‘/src’ directory changes, and another to build the ‘app’ image if the web map config SQL or GeoServer data directory changes. the ‘stack’ image needs to be built manually due to the unreliability of doing this in CI.

A vagrant base box is only created for the ‘app’ template, as this is this is the only point the application becomes useful to end-users.