Automating creating Vagrant base box version
Using a simple Python script to update the Vagrant metadata file when a new box version is produced.

As the file will be updated as part of CI we shouldn’t keep it in Git as the changes won’t be added back to the repository.

Instead we fetch the file first, add the new version (or update an existing version) and save it back.