Adding Packer support
Defines three Packer templates:

- stack (base OS, Java, Tomcat, PostgreSQL, PostGIS)
- base (GeoServer, Web Map Engine, database structure)
- app (database content, GeoServer data directory)

The stack template will be replaced with an IT image at some point. The base layer is intended to be used by specific instances of the Web Map Engine (e.g. Thwaites, ADD) to create standalone instances. The app template is an example of how to create an instance specific image.

With the exception of the stack template, GitLab CI coordinates when templates need to be rebuilt using the GitLab 'only changes' option. When built images are versioned with the current date and then stored in the BAS Repo Server. The app image is also built as a Vagrant base box for each of deployment.