Updating hardcoded WFS call for field party positions
This is equally hard-coded, but to the local development environment.

Ideally we’d be able to use the `magic.config.paths.baseurl` variable but this can’t cope with the alternate port used for GeoServer in local instances.

Instead, and anoyingly, each instance will need to modify this line to use the correct server, layer name and season.